requiremoments

+++ src/main/java/com/ii/testautomation/controllers/LicenseController.java
@@ -3,6 +3,7 @@ package com.ii.testautomation.controllers;
 import com.ii.testautomation.dto.request.LicenseRequest;
 import com.ii.testautomation.enums.RequestStatus;
 import com.ii.testautomation.response.common.BaseResponse;
+import com.ii.testautomation.service.CompanyUserService;
 import com.ii.testautomation.service.LicenseService;
 import com.ii.testautomation.utils.EndpointURI;
 import com.ii.testautomation.utils.RagexMaintainance;
@@ -20,6 +21,9 @@ public class LicenseController {
     private StatusCodeBundle statusCodeBundle;
     @Autowired
     private RagexMaintainance ragexMaintainance;
+
+    @Autowired
+    private CompanyUserService companyUserService;
 
     @PostMapping(EndpointURI.LICENSE)
     public ResponseEntity<Object> createLicense(@RequestBody LicenseRequest licenseRequest) {
@@ -50,6 +54,9 @@ public class LicenseController {
     public ResponseEntity<Object> DeleteLicense(@PathVariable Long id) {
         if (!licenseService.existsById(id)) {
             return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(), statusCodeBundle.getLicenseNotExistCode(), statusCodeBundle.getLicensePackageNotExistMessage()));
+        }
+        if (companyUserService.existsByLicenseId(id)){
+            return ResponseEntity.ok(new BaseResponse(RequestStatus.FAILURE.getStatus(),statusCodeBundle.getLicenseDeleteDependentCode(),statusCodeBundle.getLicenseDeleteDependentMessage()));
         }
         licenseService.deleteLicenseById(id);
         return ResponseEntity.ok(new BaseResponse(RequestStatus.SUCCESS.getStatus(), statusCodeBundle.getCommonSuccessCode(), statusCodeBundle.getLicenseSuccessfullyDeletedMessage()));
